Desktop/PROJETOS/DashSalaoLevel/api/requirements.txt
fastapi
uvicorn
pandas
google-auth
google-api-python-client